BrainSeg/requirements.txt
absl-py==1.3.0
astunparse==1.6.3
Bashutils==0.0.4
cachetools==5.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
exif==1.3.5
flatbuffers==22.12.6
flickrapi==2.4.0
fonttools==4.52.1
gast==0.4.0
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
hausdorff==0.2.6
idna==3.7
importlib_resources==6.4.0
itk==5.4.0
itk-core==5.4.0
itk-filtering==5.4.0
itk-io==5.4.0
itk-numerics==5.4.0
itk-registration==5.4.0
itk-segmentation==5.4.0
joblib==1.4.2
keras==2.11.0
kiwisolver==1.4.5
layers==0.1.5
libclang==14.0.6
line-profiler==4.0.2
llvmlite==0.43.0
matplotlib==3.9.0
mlxtend==0.21.0
mpmath==1.3.0
MRzeroCore @ file:///E:/MriPulse/MRTwin_pulseq/data/MRzeroCore-0.1.0-cp37-abi3-win_amd64.whl
nibabel==5.2.1
numba==0.60.0
numpy==1.23.5
oauthlib==3.2.2
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==24.0
piexif==1.1.3
pillow==10.3.0
plum-py==0.8.2
protobuf==3.19.6
pyexiv2==2.8.1
pyparsing==3.1.2
pypulseq==1.3.1.post1
python-dateutil==2.9.0.post0
PyYAML==6.0.1
requests==2.32.2
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rsa==4.9
scikit-learn==1.5.0
scipy==1.13.1
SimpleITK==2.3.1
six==1.16.0
sympy==1.12
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.29.0
termcolor==2.1.1
threadpoolctl==3.5.0
torch==2.0.1+cu117
torchaudio==2.0.2+cu117
torchkbnufft==1.3.0
torchvision==0.15.2
typing_extensions==4.12.0
umap==0.1.1
urllib3==2.2.1
wrapt==1.16.0
zipp==3.19.0
